move meshes to assets.ts
