+++ src/game/boat.ts
@@ -4,7 +4,7 @@ import { quat, vec3 } from "../gl-matrix.js";
 import { Motion, MotionDef } from "../phys_motion.js";
 import { jitter } from "../math.js";
 import { FinishedDef } from "../build.js";
-import { ColorDef, CUBE_MESH } from "./game.js";
+import { ColorDef } from "./game.js";
 import {
   MotionSmoothingDef,
   RenderableDef,
@@ -22,6 +22,7 @@ import {
 import { getAABBFromMesh, Mesh, scaleMesh3 } from "../mesh-pool.js";
 import { AABB } from "../phys_broadphase.js";
 import { Deserializer, Serializer } from "../serialize.js";
+import { CUBE_MESH } from "./assets.js";
 
 export const BoatDef = EM.defineComponent("boat", () => {
   return {
